pyproject.toml
[project]
name = "agenstin"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.42.0",
    "googlesearch-python>=1.3.0",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.3",
    "mcp>=1.26.0",
    "ollama>=0.6.1",
    "playwright>=1.58.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.0.0",
    "rich>=14.3.3",
    "slack-bolt>=1.21.0",
]
